+++ tasks.py
@@ -116,9 +116,9 @@ def set_po_metadata(filename, locale):
     with open(filename, 'rb') as infile:
         catalog = read_po(infile, locale)
     catalog.copyright_holder = 'Etalab'
-    catalog.msgid_bugs_address = 'data.gouv@data.gouv.fr'
-    catalog.language_team = 'Data.gouv.fr Team <data.gouv@data.gouv.fr>'
-    catalog.last_translator = 'Data.gouv.fr Team <data.gouv@data.gouv.fr>'
+    catalog.msgid_bugs_address = 'data.gouv@opendata.gouv.tg'
+    catalog.language_team = 'Data.gouv.fr Team <data.gouv@opendata.gouv.tg>'
+    catalog.last_translator = 'Data.gouv.fr Team <data.gouv@opendata.gouv.tg>'
     catalog.revision_date = datetime.now(LOCALTZ)
     with open(filename, 'wb') as outfile:
         write_po(outfile, catalog, width=80)
@@ -150,10 +150,10 @@ def i18n(ctx, update=False):
         ctx.run('npm run i18n:extract')
     success('Updated udata-front translations')
 
-    info('Extract data.gouv.fr-components vue translations')
+    info('Extract opendata.gouv.tg-components vue translations')
     with ctx.cd(COMPONENTS_ROOT):
         ctx.run('npm run i18n:extract')
-    success('Updated data.gouv.fr-components translations')
+    success('Updated opendata.gouv.tg-components translations')