+++ udata_front/__init__.py
@@ -1,9 +1,9 @@
 '''
-udata customizations for data.gouv.fr
+udata customizations for opendata.gouv.tg
 '''
 
 __version__ = '6.0.2.dev'
-__description__ = 'udata customizations for data.gouv.fr'
+__description__ = 'udata customizations for opendata.gouv.tg'
 
 APIGOUVFR_EXTRAS_KEY = 'apigouvfr:apis'
 APIGOUVFR_EXPECTED_FIELDS = [