+++ udata_front/theme/gouvfr/assets/js/components/dataset/CardList.vue
@@ -28,7 +28,7 @@ export type CardListProps = {
   pageSize?: number,
 };
 
-/** TODO : use a config when the component moved to data.gouv.fr-components */
+/** TODO : use a config when the component moved to opendata.gouv.tg-components */
 const defaultPageSize = 20;
 </script>