+++ udata_front/theme/gouvfr/assets/js/components/search/OrganizationSearch/OrganizationSearchOption.stories.ts
@@ -28,7 +28,7 @@ export const DefaultOrganizationSearchOption: StoryObj<typeof meta> = {
   }),
   args: {
     name: 'Etalab',
-    link: 'https://www.data.gouv.fr/fr/organizations/etalab/',
+    link: 'https://www.opendata.gouv.tg/fr/organizations/etalab/',
     logo: ''
   },
 };