+++ udata_front/theme/gouvfr/assets/js/pages/ReusePublishingForm/ReusePublishingform.stories.ts
@@ -11,7 +11,7 @@ const meta = {
 export default meta;
 
 const args = {
-  redirectDraftUrl: "https://demo.data.gouv.fr/fr/reuses/?q=draft",
+  redirectDraftUrl: "https://demo.opendata.gouv.tg/fr/reuses/?q=draft",
 };
 
 export const FormWithNoOrganizations: StoryObj<typeof meta> = {
@@ -23,7 +23,7 @@ export const FormWithNoOrganizations: StoryObj<typeof meta> = {
           first_name: "John",
           last_name: "Doe",
           roles: [],
-          avatar: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+          avatar: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
           organizations: [],
         })
       }),
@@ -35,7 +35,7 @@ export const FormWithNoOrganizations: StoryObj<typeof meta> = {
       }),
       http.post('*/api/1/reuses/:reuseId/image', async () => {
         await delay();
-        return HttpResponse.json({image: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png", success: true});
+        return HttpResponse.json({image: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png", success: true});
       }),
       http.put('*/api/1/reuses/:reuseId', async ({ request }) => {
         const body = await request.json();
@@ -66,15 +66,15 @@ export const FormWithOrganizations: StoryObj<typeof meta> = {
           first_name: "John",
           last_name: "Doe",
           roles: [],
-          avatar: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+          avatar: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
           organizations: [
             {
               acronym: null,
               badges: [],
               class: "Organization",
               id: "65e9b7cf830c3b5a515ee4ed",
-              logo: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
-              logo_thumbnail: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo_thumbnail: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
               name: "My Organization",
               page: "",
               slug: "my-organization",
@@ -85,8 +85,8 @@ export const FormWithOrganizations: StoryObj<typeof meta> = {
               badges: [],
               class: "Organization",
               id: "65e9b7cf830c3b5a515ee4ed",
-              logo: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
-              logo_thumbnail: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo_thumbnail: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
               name: "My Second Organization",
               page: "",
               slug: "my-second-organization",
@@ -103,7 +103,7 @@ export const FormWithOrganizations: StoryObj<typeof meta> = {
       }),
       http.post('*/api/1/reuses/:reuseId/image', async () => {
         await delay();
-        return HttpResponse.json({"image": "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png", "success": true});
+        return HttpResponse.json({"image": "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png", "success": true});
       }),
       http.put('*/api/1/reuses/:reuseId', async ({ request }) => {
         const body = await request.json();
@@ -134,15 +134,15 @@ export const FormWithFailedRequests: StoryObj<typeof meta> = {
           first_name: "John",
           last_name: "Doe",
           roles: [],
-          avatar: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+          avatar: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
           organizations: [
             {
               acronym: null,
               badges: [],
               class: "Organization",
               id: "65e9b7cf830c3b5a515ee4ed",
-              logo: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
-              logo_thumbnail: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo_thumbnail: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
               name: "My Organization",
               page: "",
               slug: "my-organization",
@@ -153,8 +153,8 @@ export const FormWithFailedRequests: StoryObj<typeof meta> = {
               badges: [],
               class: "Organization",
               id: "65e9b7cf830c3b5a515ee4ed",
-              logo: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
-              logo_thumbnail: "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
+              logo_thumbnail: "https://demo-static.opendata.gouv.tg/avatars/84/3194d831264f769fa817e58813d413-100.png",
               name: "My Second Organization",
               page: "",
               slug: "my-second-organization",